[Refactor] - Move Locale-related extensions to dedicated file
